requirements.txt
aniso8601==8.0.0
attrs==19.3.0
click==7.1.2
Flask==1.1.2
Flask-RESTful==0.3.8
flask-restplus==0.13.0
itsdangerous==1.1.0
Jinja2==2.11.2
jsonschema==3.2.0
MarkupSafe==1.1.1
pyrsistent==0.16.0
python-dotenv==0.14.0
pytz==2020.1
six==1.15.0
SQLAlchemy==1.3.18
Werkzeug==0.16.0
